requirements.txt
azure-identity>=1.15.0
azure-mgmt-compute>=30.0.0
azure-mgmt-resource>=23.0.1
azure-mgmt-storage>=21.0.0
boto3>=1.34.0  # For AWS integration
google-cloud-storage>=2.14.0  # For GCP integration
python-dotenv>=1.0.0
pytest>=7.4.0 